fix serialization
