+++ crates/agentgateway/src/serdes.rs
@@ -61,6 +61,32 @@ pub fn is_default<T: Default + PartialEq>(t: &T) -> bool {
 	*t == Default::default()
 }
 
+pub struct SerAsStr;
+impl<T> serde_with::SerializeAs<T> for SerAsStr
+where
+	T: AsRef<str>,
+{
+	fn serialize_as<S>(source: &T, serializer: S) -> Result<S::Ok, S::Error>
+	where
+		S: Serializer,
+	{
+		serializer.serialize_str(source.as_ref())
+	}
+}
+impl<'de, T> serde_with::DeserializeAs<'de, T> for SerAsStr
+where
+	T: std::str::FromStr,
+	<T as std::str::FromStr>::Err: Display,
+{
+	fn deserialize_as<D>(deserializer: D) -> Result<T, D::Error>
+	where
+		D: Deserializer<'de>,
+	{
+		let s = String::deserialize(deserializer)?;
+		s.parse().map_err(serde::de::Error::custom)
+	}
+}
+
 pub mod serde_instant_option {
 	use std::time::{Duration, Instant};
